test: rebuild provider mocks as scenario-based setup APIs
One mock module per provider under test/integration/mocks. Each setup
composes its own auth/identity/mailbox/calendar handlers internally and
returns a scenario API (serveCalendarEvents, rateLimitMessageList,
actAsAccount, ...), so suites never assemble raw msw handlers. A single
generic entity store replaces the duplicated label/folder stores.
