+++ packages/twenty-server/test/integration/messaging/suites/messaging-connected-account-cleanup.integration-spec.ts
@@ -2,11 +2,9 @@ import { ConnectedAccountProvider } from 'twenty-shared/types';
 import { isDefined } from 'twenty-shared/utils';
 
 import { connectMessagingAccount } from 'test/integration/messaging/utils/connect-messaging-account.util';
-import {
-  gmailMessage,
-  setupGmailMock,
-} from 'test/integration/messaging/utils/gmail-message-mock.util';
 import { deleteConnectedAccount } from 'test/integration/messaging/utils/query-messaging.util';
+import { gmailMessage } from 'test/integration/mocks/gmail-message.util';
+import { setupGoogleMock } from 'test/integration/mocks/setup-google-mock.util';
 import {
   findRecordIdsByFilter,
   findRecordNodesByFilter,
@@ -18,7 +16,7 @@ const HANDLE = 'messaging-cleanup@apple.dev';
 describe('Messaging connected account cleanup (integration)', () => {
   const inbox = [gmailMessage(), gmailMessage()];
 
-  setupGmailMock({ inbox, handle: HANDLE });
+  setupGoogleMock({ handle: HANDLE, inbox });
 
   let channel: Awaited<ReturnType<typeof connectMessagingAccount>>;
